+++ src/test/helpers/fixtures.js
@@ -1,4 +1,0 @@
-/**
- * Created by dreimus on 01/09/16.
- */
-jasmine.getFixtures().fixturesPath = "base/src/test/fixtures";